Use glib::add_idle to wakeup the gtk main loop.
This replaces busy polling. We now use a separate thread to digest the
update information from the plugins and then queue a idle_add function
that queues the gtk_redraw.

The widget needs to be in a thread local for this to work - since the
closure is run in the same thread as the gtk loop we can access it from
there this way without explicitly passing it through different threads.
I was hinted to this solution through rust-gnome/gtk#128.