Fix potential deadlock situation in the server.
Also fixes some races in the tests.

The server reregistered a socket after reading from it with MIO by
sending an event through a channel. Apparently, if during the time while
the message was queued new data arrived, MIO would not trigger. Fixed by
switching to level() instead of edge() as triggering.